feat(templates): define default ref to use for each template
